Log window extension move
