+++ docs/changes/20250731_progress.md
@@ -56,3 +56,11 @@
 ## 2025-07-19 08:46 JST [assistant]
 - simplified context creation via KafkaContextOptions.FromAppSettings
 - updated README and sample programs accordingly
+## 2025-07-19 13:44 JST [assistant]
+- moved ModelBuilderWindowExtensions to src/Core/Modeling and updated namespace
+- recorded change in diff_log/diff_window_extension_core_20250719.md
+
+
+## 2025-07-19 13:55 JST [assistant]
+- moved WindowDslExtensions into src/Core/Modeling folder as part of library core
+- verified build and tests
