+++ examples/daily-comparison/DailyComparisonLib/KafkaKsqlContext.cs
@@ -1,7 +1,7 @@
 using Kafka.Ksql.Linq.Application;
 using Kafka.Ksql.Linq.Core.Abstractions;
 using Kafka.Ksql.Linq.Core.Context;
-using Kafka.Ksql.Linq;
+using Kafka.Ksql.Linq.Core.Modeling;
 using DailyComparisonLib.Models;
 
 namespace DailyComparisonLib;
